requirements.txt
absl-py==1.4.0
annoy==1.17.2
boto3==1.26.147
botocore==1.29.147
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.10.0
filelock==3.9.0
fire==0.5.0
fonttools==4.39.4
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
idna==2.10
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.3.0
protobuf==4.23.2
psutil==5.9.5
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.0.9
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sympy==1.11.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
#torch==2.0.1+cu117
#torchaudio==2.0.2+cu117
#torchvision==0.15.2+cu117
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.114
urllib3==1.26.13
Werkzeug==2.3.4
wget==3.2
yolov5==7.0.12
fastapi==0.95.0
uvicorn[standard]==0.20.0
